Make env_helper importable from any module
